Resolve issues with ceilometer patching and packaging
Change the order that we include ceilometer-common.  This needs to
happen before we install the rest of packages because we need supporting
packages to be installed before the regular packages.

Patch ceilometer on the ceilometer-compute nodes as well as the API
nodes so we can get the database migration fixes before running a
db-sync on the compute nodes - which is completely unnecessary but
ubuntu does it for you anyway - so there's that and thanks ubuntu.
again.

Resolves Issue https://github.com/rcbops/chef-cookbooks/issues/876

(cherry picked from commit eee36100b993c53fd24c5605f28e23e5c4e05d67)