Added to fix RHEL specific issues with config files
When installing on RHEL or using the RDO packages
REDHAT installs configuration files for specific services
into directories under `/usr/share` these additional configuration
files are in direct conflict with the configs that are in `/etc/`

To fix this issue we are truncating the file `ceilometer-dist.conf` and
overloading the file in ceilometer.conf with the values that were in the
dist.conf file.  In this case we are simply setting the control_exchange
and which RPC driver to use.